Do not allow to suggest duplicate option
